requirements.txt
Flask==1.1.2
python-dateutil==2.8.1
opencv-python==4.5.1.48
imutils==0.5.4
numpy==1.20.1
pytesseract==0.3.7